+++ cmd/veneur-prometheus/main.go
@@ -6,6 +6,8 @@ import (
 	"io"
 	"math"
 	"net/http"
+	"regexp"
+	"strings"
 	"time"
 
 	"github.com/DataDog/datadog-go/statsd"
@@ -15,11 +17,13 @@ import (
 )
 
 var (
-	debug       = flag.Bool("d", false, "Enable debug mode")
-	metricsHost = flag.String("h", "http://localhost:9090/metrics", "The full URL — like 'http://localhost:9090/metrics' to query for Prometheus metrics.")
-	interval    = flag.String("i", "10s", "The interval at which to query. Value must be parseable by time.ParseDuration (https://golang.org/pkg/time/#ParseDuration).")
-	prefix      = flag.String("p", "", "A prefix to append to any metrics emitted. Do not include a trailing period.")
-	statsHost   = flag.String("s", "127.0.0.1:8126", "The host and port — like '127.0.0.1:8126' — to send our metrics to.")
+	debug             = flag.Bool("d", false, "Enable debug mode")
+	metricsHost       = flag.String("h", "http://localhost:9090/metrics", "The full URL — like 'http://localhost:9090/metrics' to query for Prometheus metrics.")
+	interval          = flag.String("i", "10s", "The interval at which to query. Value must be parseable by time.ParseDuration (https://golang.org/pkg/time/#ParseDuration).")
+	ignoredLabelsStr  = flag.String("ignored-labels", "", "A comma-seperated list of label name regexes to not export")
+	ignoredMetricsStr = flag.String("ignored-metrics", "", "A comma-seperated list of metric name regexes to not export")
+	prefix            = flag.String("p", "", "A prefix to append to any metrics emitted. Do not include a trailing period.")
+	statsHost         = flag.String("s", "127.0.0.1:8126", "The host and port — like '127.0.0.1:8126' — to send our metrics to.")
 )
 
 func main() {
@@ -36,9 +40,19 @@ func main() {
 		logrus.WithError(err).Fatalf("Failed to parse interval '%s'", *interval)
 	}
 
+	ignoredLabels := []*regexp.Regexp{}
+	ignoredMetrics := []*regexp.Regexp{}
+
+	for _, ignoredLabelStr := range strings.Split(*ignoredLabelsStr, ",") {
+		ignoredLabels = append(ignoredLabels, regexp.MustCompile(ignoredLabelStr))
+	}
+	for _, ignoredMetricStr := range strings.Split(*ignoredMetricsStr, ",") {
+		ignoredMetrics = append(ignoredLabels, regexp.MustCompile(ignoredMetricStr))
+	}
+
 	ticker := time.NewTicker(i)
 	for _ = range ticker.C {
-		collect(c)
+		collect(c, ignoredLabels, ignoredMetrics)
 	}
 
 	if *prefix != "" {
@@ -46,7 +60,7 @@ func main() {
 	}
 }
 
-func collect(c *statsd.Client) {
+func collect(c *statsd.Client, ignoredLabels []*regexp.Regexp, ignoredMetrics []*regexp.Regexp) {
 	logrus.WithFields(logrus.Fields{
 		"stats_host":   *statsHost,
 		"metrics_host": *metricsHost,
@@ -66,35 +80,27 @@ func collect(c *statsd.Client) {
 			break
 		}
 
+		if !shouldExportMetric(mf, ignoredMetrics) {
+			continue
+		}
+
 		var metricCount int64
 		switch mf.GetType() {
 		case dto.MetricType_COUNTER:
 			for _, counter := range mf.GetMetric() {
-				var tags []string
-				labels := counter.GetLabel()
-				for _, pair := range labels {
-					tags = append(tags, fmt.Sprintf("%s:%s", pair.GetName(), pair.GetValue()))
-				}
+				tags := getTags(counter.GetLabel(), ignoredLabels)
 				c.Count(mf.GetName(), int64(counter.GetCounter().GetValue()), tags, 1.0)
 				metricCount++
 			}
 		case dto.MetricType_GAUGE:
 			for _, gauge := range mf.GetMetric() {
-				var tags []string
-				labels := gauge.GetLabel()
-				for _, pair := range labels {
-					tags = append(tags, fmt.Sprintf("%s:%s", pair.GetName(), pair.GetValue()))
-				}
+				tags := getTags(gauge.GetLabel(), ignoredLabels)
 				c.Gauge(mf.GetName(), float64(gauge.GetGauge().GetValue()), tags, 1.0)
 				metricCount++
 			}
 		case dto.MetricType_HISTOGRAM, dto.MetricType_SUMMARY:
 			for _, histo := range mf.GetMetric() {
-				var tags []string
-				labels := histo.GetLabel()
-				for _, pair := range labels {
-					tags = append(tags, fmt.Sprintf("%s:%s", pair.GetName(), pair.GetValue()))
-				}
+				tags := getTags(histo.GetLabel(), ignoredLabels)
 				hname := mf.GetName()
 				summ := histo.GetSummary()
 				c.Gauge(fmt.Sprintf("%s.sum", hname), summ.GetSampleSum(), tags, 1.0)
@@ -112,4 +118,39 @@ func collect(c *statsd.Client) {
 		}
 		c.Count("veneur.prometheus.metrics_flushed_total", metricCount, nil, 1.0)
 	}
+}
+
+func getTags(labels []*dto.LabelPair, ignoredLabels []*regexp.Regexp) []string {
+	var tags []string
+
+	for _, pair := range labels {
+		labelName := pair.GetName()
+		labelValue := pair.GetValue()
+		include := true
+
+		for _, ignoredLabel := range ignoredLabels {
+			if ignoredLabel.MatchString(labelName) {
+				include = false
+				break
+			}
+		}
+
+		if include {
+			tags = append(tags, fmt.Sprintf("%s:%s", labelName, labelValue))
+		}
+	}
+
+	return tags
+}
+
+func shouldExportMetric(mf dto.MetricFamily, ignoredMetrics []*regexp.Regexp) bool {
+	for _, ignoredMetric := range ignoredMetrics {
+		metricName := mf.GetName()
+
+		if ignoredMetric.MatchString(metricName) {
+			return false
+		}
+	}
+
+	return true
 }